Multi-target support
The target/platforms specific code was moved to the multitargetsupport.h
header file, so all the  duplicate code is there now.

Added the Visual Studio solution and project files.

Fine tuned the .gitignore file to reflect the changes made by Visual
Studio.

The codebase is now nearly Visual C compliant, the only thing which is
not working is the flywheel control.

The compilation was tested with Visual Studio 2013 Community Edition.

During this refactoring none of the old functions were affected, the Qt
compilation is the same as before.